Move custom widgets from pdf app to chords/widgets.py
