+++ pdf/forms.py
@@ -9,53 +9,12 @@ from django.forms import (
     Form,
     IntegerField,
     ModelMultipleChoiceField,
-    Widget,
-    TextInput,
 )
-from django.utils.safestring import mark_safe
 from django.utils.translation import gettext_lazy as _
 
 from backend.models import Song
+from chords.widgets import NoopWidget, AppendWidget
 from pdf.models.request import PDFSong, ManualPDFTemplate, PDFFile
-
-
-class NoopWidget(Widget):
-    """
-    Renders completely empty widget, but still retrieves data
-    Useful for situation when you handle submit specifically
-    """
-
-    def render(self, name, value, attrs=None, renderer=None):
-        return ""
-
-    def value_from_datadict(self, data, files, name):
-        try:
-            getter = data.getlist
-        except AttributeError:
-            getter = data.get
-        return getter(name)
-
-    @property
-    def is_hidden(self):
-        return True
-
-
-class PrependWidget(TextInput):
-    """Widget that prepend boostrap-style span with data to specified base widget"""
-
-    def __init__(self, data, *args, **kwargs):
-        """Initialise widget and get base instance"""
-        super().__init__(*args, **kwargs)
-        self.data = data
-
-    def render(self, name, value, attrs=None, renderer=None):
-        """Render base widget and add bootstrap spans"""
-        context = self.get_context(name, value, attrs)
-        field = self._render(self.template_name, context, renderer)
-        return mark_safe(
-            ('<div class="input-group mb-3">' "  %(field)s" '  <span class="input-group-text">%(data)s</span>' "</div>")
-            % {"field": field, "data": self.data}
-        )
 
 
 class ManualTemplateForm(ModelForm):
@@ -64,13 +23,13 @@ class ManualTemplateForm(ModelForm):
     class Meta:
         model = ManualPDFTemplate
         fields = ["name", "display_name", "filename", "title", "image", "public", "locale", "show_date", "link"]
-        widgets = {"filename": PrependWidget(".pdf")}
+        widgets = {"filename": AppendWidget(".pdf")}
 
 
 class PDFFileEditForm(ModelForm):
     """Edit form for PDFFile, allows changing only the display name"""
 
-    filename = CharField(widget=PrependWidget(".pdf"), disabled=True)
+    filename = CharField(widget=AppendWidget(".pdf"), disabled=True)
 
     class Meta:
         model = PDFFile
